replace TestUtils class with TestHelpers
